src/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic-settings==2.2.1
aiofiles==23.2.1
langchain==0.1.20
pyMuPDF==1.24.3
motor==3.6.0
openai==1.75.0
cohere==4.45.0
groq==0.33.0
qdrant-client==1.10.0
SQLAlchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2==2.9.10
pgvector==0.4.0
nltk==3.9.1

# Monitoring and Metrics
prometheus-client==0.19.0
starlette-exporter==0.17.1
fastapi-health==0.4.0

# Task Queue and Background Processing
celery==5.5.3
redis==6.2.0
kombu==5.5.4
billiard==4.2.1
vine==5.1.0
flower==2.0.1
